Mover StarFieldAdaptor to STARField.h
