+++ StRoot/StFwdTrackMaker/StFwdTrackMaker.cxx
@@ -174,36 +174,6 @@ class SiRasterizer {
     double raster_r, raster_phi;
 };
 
-//_______________________________________________________________________________________
-// Adaptor for STAR magnetic field
-class StarFieldAdaptor : public genfit::AbsBField {
-  public:
-    StarFieldAdaptor() {
-        _gField = this; // global pointer in TrackFitter.h
-    };
-    virtual TVector3 get(const TVector3 &position) const {
-        double x[] = {position[0], position[1], position[2]};
-        double B[] = {0, 0, 0};
-
-        if (StarMagField::Instance())
-            StarMagField::Instance()->Field(x, B);
-
-        return TVector3(B);
-    };
-    virtual void get(const double &_x, const double &_y, const double &_z, double &Bx, double &By, double &Bz) const {
-        double x[] = {_x, _y, _z};
-        double B[] = {0, 0, 0};
-
-        if (StarMagField::Instance())
-            StarMagField::Instance()->Field(x, B);
-
-        Bx = B[0];
-        By = B[1];
-        Bz = B[2];
-        return;
-    };
-};
-
 //  Wrapper class around the forward tracker
 class ForwardTracker : public ForwardTrackMaker {
   public:
